moved nav component
