+++ src/components/layout/index.tsx
@@ -7,7 +7,7 @@
 
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
-import Nav from "../Nav"
+import Nav from "../specific/Nav"
 import Header from "../Header"
 import ScrollToTop from "react-scroll-up"
 import styles from "./styles.module.scss"
